requirements.txt
django==3.0.5
pillow==7.1.2
django-pure-pagination==0.3.0
redis==3.5.1
requests==2.23.0
future==0.18.2
django-crispy-forms==1.9.1
six==1.15.0
django-formtools==2.2
django-import-export==2.1.0
httplib2==0.18.1
django-simple-captcha==0.5.12
mysqlclient==1.4.6